+++ contracts/l1/L1Escrow.sol
@@ -1,4 +1,5 @@
 // SPDX-License-Identifier: UNLICENSED
+
 pragma solidity ^0.7.6;
 
 interface ApproveLike {
@@ -9,7 +10,7 @@ interface ApproveLike {
 contract L1Escrow {
 
   // --- Auth ---
-  mapping (address => uint8) public wards;
+  mapping (address => uint256) public wards;
   function rely(address usr) external auth {
     wards[usr] = 1;
     emit Rely(usr);
